requirements.txt
beautifulsoup4==4.12.3
numpy==1.24.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
Requests==2.31.0
streamlit==1.34.0
supervision==0.20.0
torch==2.3.0
ultralytics==8.2.15
